feat(tsp): added progess-bar and new tsp more efficient

+++ application/src/main/java/fr/delivrooom/application/model/tsp/TSP1.java
@@ -1,6 +1,8 @@
 package fr.delivrooom.application.model.tsp;
 
+import fr.delivrooom.application.model.DeliveriesDemand;
 import fr.delivrooom.application.model.Graphe;
+import fr.delivrooom.application.port.out.NotifyTSPProgressToGui;
 
 import java.util.Collection;
 import java.util.Iterator;
@@ -8,7 +10,7 @@ import java.util.Iterator;
 public class TSP1 extends TemplateTSP {
 
     @Override
-    protected int bound(Long sommetCourant, Collection<Long> nonVus) {
+    protected float bound(Long sommetCourant, Collection<Long> nonVus) {
         return 0;
     }
 
@@ -16,10 +18,5 @@ public class TSP1 extends TemplateTSP {
     protected Iterator<Long> iterator(Long sommetCrt, Collection<Long> nonVus, Graphe g) {
         return new IteratorSeq(nonVus, sommetCrt, g);
     }
-
-
-
-
-
 
 }
